+++ src/Services/Catalog/Product/Sku/Result/SkusResult.php
@@ -3,7 +3,7 @@
 /**
  * This file is part of the bitrix24-php-sdk package.
  *
- * © Maksim Mesilov <mesilov.maxim@gmail.com>
+ * © Dmitriy Ignatenko <algonexys@gmail.com>
  *
  * For the full copyright and license information, please view the MIT-LICENSE.txt
  * file that was distributed with this source code.
@@ -11,22 +11,22 @@
 
 declare(strict_types=1);
 
-namespace Bitrix24\SDK\Services\Catalog\Catalog\Result;
+namespace Bitrix24\SDK\Services\Catalog\Product\Sku\Result;
 
 use Bitrix24\SDK\Core\Exceptions\BaseException;
 use Bitrix24\SDK\Core\Result\AbstractResult;
 
-class CatalogsResult extends AbstractResult
+class SkusResult extends AbstractResult
 {
     /**
-     * @return CatalogItemResult[]
+     * @return SkuItemResult[]
      * @throws BaseException
      */
-    public function getCatalogs(): array
+    public function getSkus(): array
     {
         $res = [];
-        foreach ($this->getCoreResponse()->getResponseData()->getResult()['catalogs'] as $catalog) {
-            $res[] = new CatalogItemResult($catalog);
+        foreach ($this->getCoreResponse()->getResponseData()->getResult()['units'] as $unit) {
+            $res[] = new SkuItemResult($unit);
         }
 
         return $res;